test(gepay): corriger GePayMtnTokenHeaderTest et cast float PartnerFinancialDashboard
GePayMtnTokenHeaderTest : après le fix P0 (référence MTN persistée avant appel
réseau), MtnMomoProvider::initiate() écrit maintenant dans gepay_transactions
avant l'appel HTTP. Le test utilisait un modèle non persisté (id=null) et aucune
migration — ajout de RefreshDatabase + création du GePayClient + GePayTransaction
en base pour que persistProviderReference() s'exécute normalement.

PartnerFinancialDashboardService : max($net - $alreadyPaid - $pendingPayouts, 0)
retourne 0 (int) quand 0 gagne — cast (float) pour que assertSame(0.0, ...) passe.

765/765 tests verts.

+++ app/Services/PartnerFinancialDashboardService.php
@@ -126,8 +126,8 @@ class PartnerFinancialDashboardService
         float $pendingPayouts,
         array $overrides = []
     ): array {
-        $net = max($gross - $commission, 0);
-        $available = max($net - $alreadyPaid - $pendingPayouts, 0);
+        $net = (float) max($gross - $commission, 0);
+        $available = (float) max($net - $alreadyPaid - $pendingPayouts, 0);
 
         $cards = [
             [
